+++ src/document.h
@@ -30,7 +30,7 @@ extern uint8_t create_default_ruler(uint16_t ruler_addr);
 
 extern uint8_t get_byte_from_file(void);
 
-extern uint16_t* get_register_address(uint8_t a);
+extern unsigned int* get_register_address(uint8_t a);
 
 extern uint8_t initialise_document(void);
